Setup ETL project for ETL operations
Instead of using complex tool, we code the data manipulations in C# instead to test the waters.
It may not be 100% automated but it provides greater control on the data transformation and the change tracking (via git).
Note: we target .netcore for this project because the ETLBox nuget package target this framework as well and also because it should be deployable more easily later (than the .net framework)